requirements.txt
lxml==5.3.0
python-dateutil==2.9.0.post0
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pytest==8.3.3
